Only allow awesome table block in news pages
